chore(Bonus Pagamenti Digitali): [#175589933] Sync PM/walletV2 spec (#2342)
* [#175589933] refactoring

* [#175589933] refactoring

* [#175589933] refactoring

* [#175589933] fix and refactoring

* [#175589933] update comment

* [#175589933] update comment

* [#175589933] refactoring

* [#175589933] refactoring

* [#175589933] refactoring

* [#175589933] fix lints errors
