+++ package.json
@@ -10,7 +10,7 @@
   "io_bpd_award_periods": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/bonus/specs/bpd/award.json",
   "io_bpd_winning_transactions": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/bonus/specs/bpd/winning_transactions.json",
   "pagopa_api": "https://wisp2.pagopa.gov.it/pp-restapi-CD/v2/api-docs",
-  "pagopa_api_bancomat": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/bonus/specs/bpd/pm/swagger_bancomat.json",
+  "pagopa_api_walletv2": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/bonus/specs/bpd/pm/walletv2.json",
   "private": true,
   "scripts": {
     "start": "react-native start",
@@ -41,12 +41,12 @@
     "generate:bpd-payment": "rimraf definitions/bpd/payment && mkdir -p definitions/bpd/payment  && gen-api-models --api-spec $npm_package_io_bpd_payment --out-dir ./definitions/bpd/payment --no-strict --request-types --response-decoders",
     "generate:bpd-award-periods": "rimraf definitions/bpd/award_periods && mkdir -p definitions/bpd/award_periods  && gen-api-models --api-spec $npm_package_io_bpd_award_periods --out-dir ./definitions/bpd/award_periods --no-strict --request-types --response-decoders",
     "generate:bpd-winning-transactions": "rimraf definitions/bpd/winning_transactions && mkdir -p definitions/bpd/winning_transactions  && gen-api-models --api-spec $npm_package_io_bpd_winning_transactions --out-dir ./definitions/bpd/winning_transactions --no-strict --request-types --response-decoders",
-    "generate:pagopa-api-bancomat": "rimraf definitions/pagopa/bancomat && mkdir -p definitions/pagopa/bancomat  && gen-api-models --api-spec $npm_package_pagopa_api_bancomat --out-dir ./definitions/pagopa/bancomat --no-strict --request-types --response-decoders",
+    "generate:pagopa-api-walletv2": "rimraf definitions/pagopa/walletv2 && mkdir -p definitions/pagopa/walletv2  && gen-api-models --api-spec $npm_package_pagopa_api_walletv2 --out-dir ./definitions/pagopa/walletv2 --no-strict --request-types --response-decoders",
     "bundle:android-release": "node node_modules/react-native/local-cli/cli.js bundle --platform android --dev false --entry-file index.js --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res/",
     "generate:locales": "ts-node --skip-project -O '{\"lib\":[\"es2015\"]}' scripts/make-locales.ts",
     "generate:content-definitions": "rimraf definitions/content && mkdir -p definitions/content && gen-api-models --api-spec $npm_package_io_content_specs --out-dir ./definitions/content",
     "generate:bpd": "npm-run-all generate:bpd-citizen generate:bpd-payment generate:bpd-award-periods generate:bpd-winning-transactions",
-    "generate:all": "npm-run-all generate:mock-google-services-json generate:api-definitions generate:pagopa-api-definitions generate:pagopa-api-bancomat generate:content-definitions generate:bonus-vacanze-definitions generate:bpd generate:locales",
+    "generate:all": "npm-run-all generate:mock-google-services-json generate:api-definitions generate:pagopa-api-definitions generate:pagopa-api-walletv2 generate:content-definitions generate:bonus-vacanze-definitions generate:bpd generate:locales",
     "locales_unused": "ts-node --skip-project -O '{\"lib\":[\"es2015\"]}' scripts/unused-locales.ts",
     "generate-myportal-readme": "jsdoc2md ts/utils/webviewScripts/*.js > MYPORTAL_README.md"
   },
