+++ ts/features/wallet/onboarding/bancomat/components/BankPreviewItem.tsx
@@ -2,7 +2,7 @@ import { ListItem, View } from "native-base";
 import * as React from "react";
 import { Image, ImageStyle, StyleProp, StyleSheet } from "react-native";
 import I18n from "../../../../../i18n";
-import { Abi } from "../../../../../../definitions/pagopa/bancomat/Abi";
+import { Abi } from "../../../../../../definitions/pagopa/walletv2/Abi";
 import ButtonDefaultOpacity from "../../../../../components/ButtonDefaultOpacity";
 import { LabelSmall } from "../../../../../components/core/typography/LabelSmall";
 import { IOColors } from "../../../../../components/core/variables/IOColors";
