--- ts/features/wallet/onboarding/bancomat/mock/mockData.ts
@@ -1,3037 +1,0 @@
-import { Abi } from "../../../../../../definitions/pagopa/bancomat/Abi";
-import {
-  Card,
-  ProductTypeEnum,
-  ValidityStateEnum
-} from "../../../../../../definitions/pagopa/bancomat/Card";
-
-export const abis: ReadonlyArray<Abi> = [
-  {
-    abi: "00001",
-    name: "Dutto Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00002",
-    name: "Mignogna - Repetto"
-  },
-  {
-    abi: "00003",
-    name: "Brambilla, Schiavon and Brogi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00004",
-    name: "Rocca e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00005",
-    name: "Mereu, Carrieri and Casadei",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00006",
-    name: "Castiglioni - Panzeri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00007",
-    name: "Agostinelli - Nocentini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00008",
-    name: "Melandri, Zennaro and Venturini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00009",
-    name: "Bianco e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00010",
-    name: "Martinelli - Lodi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00011",
-    name: "Semeraro, Di Dio and Conte",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00012",
-    name: "Carta SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00013",
-    name: "Surace - Di Paolo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00014",
-    name: "Manetti, Pepe and Carletti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00015",
-    name: "Repetto e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00016",
-    name: "Lodi, Di Girolamo and Parodi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00017",
-    name: "Ceccarelli, Proietti and Marangon",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00018",
-    name: "Fontana SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00019",
-    name: "Ercoli SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00020",
-    name: "Fiorini s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00021",
-    name: "Demurtas Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00022",
-    name: "Ferrara, Favre and Tripodi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00023",
-    name: "Cerrani s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00024",
-    name: "Ciuffreda SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00025",
-    name: "Moretti - Ferrando",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00026",
-    name: "Paolicelli, Dalmasso and Tesi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00027",
-    name: "Serafini - Marelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00028",
-    name: "Montanari, Pozzo and Conte",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00029",
-    name: "Ferraro SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00030",
-    name: "Rota - Coviello",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00031",
-    name: "Iezzi, Gentile and Gallo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00032",
-    name: "Mazza - Fiorini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00033",
-    name: "Sartor, Cortinovis and Brumat",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00034",
-    name: "Lai - Pellegrino",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00035",
-    name: "Delfino - Mancuso",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00036",
-    name: "Maggiani - Battistini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00037",
-    name: "Penzo SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00038",
-    name: "Careddu, Tallarico and Sorrentino",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00039",
-    name: "Corridori e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00040",
-    name: "Mauro e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00041",
-    name: "Sabatini, Marini and Bianchi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00042",
-    name: "Quinto - Franco",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00043",
-    name: "Fanelli s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00044",
-    name: "Miglio SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00045",
-    name: "Martino, Di Dio and Rossi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00046",
-    name: "De Col Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00047",
-    name: "Sbrana SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00048",
-    name: "Pellegrini Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00049",
-    name: "Fiore SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00050",
-    name: "Arena s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00051",
-    name: "Santini e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00052",
-    name: "Coslovich e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00053",
-    name: "Pellegrino - Barbieri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00054",
-    name: "Pintus SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00055",
-    name: "Gargiulo, Battaglia and Innocenti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00056",
-    name: "Pellegrini e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00057",
-    name: "Di Girolamo - Costantini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00058",
-    name: "Carrieri, Costantini and Pardini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00059",
-    name: "Marziali Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00060",
-    name: "Valentini, Picco and Vuillermoz",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00061",
-    name: "Cerutti - Cavallo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00062",
-    name: "Venturini SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00063",
-    name: "Dutto - Ruggeri",
-    logoUrl:
-      "https://www.symphonia.it/sites/default/files/Logo%20Bim%20Verticale%20NEW_ritagliato.png?fid=2631"
-  },
-  {
-    abi: "00064",
-    name: "Ferrero - Garau",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00065",
-    name: "Filippi, Giunta and Pession",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00066",
-    name: "Tomasi s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00067",
-    name: "Donda - Zanella",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00068",
-    name: "Brambilla, Rodigari and Romeo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00069",
-    name: "Sala Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00070",
-    name: "Franzè s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00071",
-    name: "Coviello, Casadio and Cascone",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00072",
-    name: "Franzè - Martini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00073",
-    name: "Senatore, Marangon and Romaniello",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00074",
-    name: "Melandri, Carletti and Coppola",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00075",
-    name: "Franco Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00076",
-    name: "Blanc e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00077",
-    name: "Gallo - Grandi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00078",
-    name: "Corazza s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00079",
-    name: "Luciani - Rodigari",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00080",
-    name: "Sartor - Balboni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00081",
-    name: "Demurtas e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00082",
-    name: "Surace - Martini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00083",
-    name: "Furlan - Crevatin",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00084",
-    name: "Gambino, Grange and Martino",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00085",
-    name: "Semeraro Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00086",
-    name: "Greppi - Gazzola",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00087",
-    name: "Pozzo - Pappalardo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00088",
-    name: "Prencipe e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00089",
-    name: "Rinaldi - Casadei",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00090",
-    name: "De Bona s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00091",
-    name: "Moro SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00092",
-    name: "Rota Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00093",
-    name: "Matera - Manca",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00094",
-    name: "Vierin s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00095",
-    name: "Lombardi - Frigo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00096",
-    name: "Penzo - Torchio",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00097",
-    name: "Sorrentino - Sucera",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00098",
-    name: "Ramella - Romaniello",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00099",
-    name: "Fabris SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00100",
-    name: "Manetti - Carini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00101",
-    name: "Gagliano - Cossu",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00102",
-    name: "Mastroianni Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00103",
-    name: "Bartolucci s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00104",
-    name: "Melandri, Grillo and Leone",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00105",
-    name: "Lodi, Dal Farra and Di Felice",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00106",
-    name: "Fabbri, Cattaneo and Campisi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00107",
-    name: "Guerrini e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00108",
-    name: "Campisi s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00109",
-    name: "Barbera, Carvelli and Marelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00110",
-    name: "Leonetti - Manetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00111",
-    name: "De Lucia, Ruggiero and Gregori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00112",
-    name: "Cappello Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00113",
-    name: "Coslovich Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00114",
-    name: "Pinna e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00115",
-    name: "Zani, Vezzoli and Gori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00116",
-    name: "Gabrielli - Franzoni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00117",
-    name: "Perego - Benini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00118",
-    name: "Brambilla s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00119",
-    name: "Vitale, Foschi and Laganà",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00120",
-    name: "Giordano - Carlucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00121",
-    name: "Mosca, Barsotti and Mereu",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00122",
-    name: "Romaniello Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00123",
-    name: "Fiore, Sanfilippo and Epifani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00124",
-    name: "Luciani e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00125",
-    name: "Montanari - Guastella",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00126",
-    name: "Mignogna, Sorrentino and Palmisano",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00127",
-    name: "Ravaioli, Giovannini and Lanteri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00128",
-    name: "Gatti, Gambino and Di Paolo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00129",
-    name: "Pedrotti Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00130",
-    name: "Guastone - Bartolini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00131",
-    name: "Rodigari - Pepe",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00132",
-    name: "Tagliabue, Cerutti and Lanza",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00133",
-    name: "Rea, Barsotti and Giannini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00134",
-    name: "Ferro Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00135",
-    name: "Cerrani s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00136",
-    name: "Di Matteo e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00137",
-    name: "Agostinelli, Ferluga and Molteni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00138",
-    name: "Sorrentino e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00139",
-    name: "Martino, Pepe and D'Andrea",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00140",
-    name: "Kofler - Favre",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00141",
-    name: "Foschi - Natale",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00142",
-    name: "Cozzani - Spano",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00143",
-    name: "Silenzi, Canova and Casini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00144",
-    name: "Nobili e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00145",
-    name: "Bodini - Carnovale",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00146",
-    name: "Tagliabue s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00147",
-    name: "Bocchi s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00148",
-    name: "Ruggiero, Beltrami and Francese",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00149",
-    name: "Quarta Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00150",
-    name: "Bodini - Locatelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00151",
-    name: "Masiero - Di Nardo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00152",
-    name: "Spina e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00153",
-    name: "Leonetti Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00154",
-    name: "Corti e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00155",
-    name: "Piazzolla, Furlan and Ciavarella",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00156",
-    name: "Ruggiero - Di Carlo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00157",
-    name: "Guerra, Di Francesco and Rizzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00158",
-    name: "Carvelli s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00159",
-    name: "Sabbatini, Crea and Calligaris",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00160",
-    name: "Lonardi - Greppi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00161",
-    name: "Fontana Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00162",
-    name: "Maggi, Sacco and Massaro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00163",
-    name: "Menegatti s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00164",
-    name: "Foti, Gruber and Spano",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00165",
-    name: "Dragoni - Giovannini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00166",
-    name: "Di Cesare SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00167",
-    name: "Granata - Fasoli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00168",
-    name: "Baggio, Romagnoli and Gregori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00169",
-    name: "Locatelli - Grange",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00170",
-    name: "Randazzo, Pace and Rodigari",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00171",
-    name: "Venturi - Amerio",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00172",
-    name: "Battaglia - D'Amico",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00173",
-    name: "Torchio - Manca",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00174",
-    name: "Manetti Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00175",
-    name: "Mignogna e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00176",
-    name: "Tassone, Gentili and Fonda",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00177",
-    name: "Bernardini - Melis",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00178",
-    name: "Grange, Montemurro and Minniti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00179",
-    name: "Ruggeri, Angelucci and Innocenti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00180",
-    name: "Massaro - Zennaro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00181",
-    name: "Crespi SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00182",
-    name: "Caputo, Egger and Beltrami",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00183",
-    name: "Di Girolamo - Tagliabue",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00184",
-    name: "Cozzani, Battaglia and Vichi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00185",
-    name: "Murgia - Licitra",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00186",
-    name: "D'Amico SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00187",
-    name: "Puglisi - Marega",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00188",
-    name: "Barbieri, Galati and Lai",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00189",
-    name: "Briano - Morelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00190",
-    name: "De Luca, Rodigari and Bionaz",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00191",
-    name: "Montagna - Torchio",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00192",
-    name: "Luciani - Pieri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00193",
-    name: "Bernasconi, Di Paolo and Grandi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00194",
-    name: "Scalise Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00195",
-    name: "Rosati - Mora",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00196",
-    name: "Scarpa, Menna and Piazzolla",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00197",
-    name: "Salvucci - Murgia",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00198",
-    name: "Diana Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00199",
-    name: "Martinelli - Occhipinti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00200",
-    name: "Di Iorio, Amato and Bertolussi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00201",
-    name: "Meloni - Rosset",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00202",
-    name: "Moffa - Cozzani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00203",
-    name: "Riggi, Forte and Barbieri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00204",
-    name: "Trevisan - Sechi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00205",
-    name: "Degrassi SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00206",
-    name: "Luppi - Sbrana",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00207",
-    name: "Galati s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00208",
-    name: "Pastorino, Tiozzo and Schiavon",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00209",
-    name: "Barsotti - Ferraris",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00210",
-    name: "Zennaro, Traverso and Schiavon",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00211",
-    name: "Montalbano - Azara",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00212",
-    name: "Benvenuti s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00213",
-    name: "Berto, Pieri and Scherini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00214",
-    name: "Gori SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00215",
-    name: "Luciani, Pardini and Di Marco",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00216",
-    name: "Vuillermoz - Crippa",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00217",
-    name: "Manetti - Briano",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00218",
-    name: "Mereu s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00219",
-    name: "Ferraro - Piscitelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00220",
-    name: "Spina - Righetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00221",
-    name: "Alaimo, Cossu and Parodi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00222",
-    name: "Bonato, Cerutti and Catellani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00223",
-    name: "Ciuffreda Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00224",
-    name: "Menozzi, Tanzi and Corridori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00225",
-    name: "Pession s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00226",
-    name: "Lorusso - Iannaccone",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00227",
-    name: "Canova - Brogi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00228",
-    name: "Ferretti, Pivetta and Pinna",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00229",
-    name: "Pinna e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00230",
-    name: "Gurrieri - Rossi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00231",
-    name: "Lodi - Vuillermoz",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00232",
-    name: "Monaci - Aiello",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00233",
-    name: "Di Francesco, Tallarico and Merola",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00234",
-    name: "Favaro - Dal Farra",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00235",
-    name: "Cecchini s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00236",
-    name: "Furlan e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00237",
-    name: "Mosti e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00238",
-    name: "Cerise e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00239",
-    name: "Bruno e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00240",
-    name: "Leonetti, Perri and Martino",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00241",
-    name: "Corradi - Sartori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00242",
-    name: "Giunta, Vincenzi and Rossi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00243",
-    name: "Gazzola Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00244",
-    name: "Riggi, De Giorgi and Nanni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00245",
-    name: "Baldi - Canu",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00246",
-    name: "Marega, Garavaglia and Mazzeo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00247",
-    name: "Minniti Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00248",
-    name: "Vitali, Serra and Palmisano",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00249",
-    name: "Beltrame - Macchi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00250",
-    name: "Fumagalli, Pavan and Gentili",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00251",
-    name: "Scarpa e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00252",
-    name: "Franzoni, Iezzi and Milani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00253",
-    name: "Maggiani - Alaimo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00254",
-    name: "Riva e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00255",
-    name: "Penzo - Gargiulo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00256",
-    name: "Bortolin, Montalbano and Montanari",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00257",
-    name: "Murru, Marconi and Capasso",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00258",
-    name: "Ferro, Crea and Rizzo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00259",
-    name: "Zennaro, Tesi and Grange",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00260",
-    name: "Martinelli, Marinelli and Cavallo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00261",
-    name: "Ravaioli, Giunta and Parente",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00262",
-    name: "Lombardo, Tallarico and Casadei",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00263",
-    name: "Licata, Rota and Gentile",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00264",
-    name: "Romeo - Ciuffreda",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00265",
-    name: "De Angelis e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00266",
-    name: "Hofer s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00267",
-    name: "Di Giuseppe, Di Paolo and Mereu",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00268",
-    name: "Ciavarella, Mura and Morandi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00269",
-    name: "Marelli - Bertolussi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00270",
-    name: "Mazza Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00271",
-    name: "Iotti SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00272",
-    name: "Nocentini s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00273",
-    name: "Merola s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00274",
-    name: "Colonna Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00275",
-    name: "Gori SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00276",
-    name: "Guerriero, Fumarola and Rizzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00277",
-    name: "Pini s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00278",
-    name: "Venezia, Gurrieri and Moser",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00279",
-    name: "Carrieri s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00280",
-    name: "Bisio, Azara and Azara",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00281",
-    name: "Bortolin - Locatelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00282",
-    name: "Pappalardo e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00283",
-    name: "Bartolini, Gabriele and Mosca",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00284",
-    name: "Caruso, Penzo and Calligaris",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00285",
-    name: "Gasser e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00286",
-    name: "Marchetti, Bianchi and Atzeni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00287",
-    name: "Martini, Ferrero and Melis",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00288",
-    name: "Bocchi - Silenzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00289",
-    name: "Baggio, Tassone and Napolitano",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00290",
-    name: "Di Domenico Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00291",
-    name: "Melandri, Davoli and Bucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00292",
-    name: "Ferrario - Cavalli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00293",
-    name: "Venezia SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00294",
-    name: "Pavan - Veronese",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00295",
-    name: "Pacini - Righi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00296",
-    name: "Pellegrino - Dalmasso",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00297",
-    name: "Menegatti - Corridori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00298",
-    name: "Zanatta s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00299",
-    name: "Viale, Favre and Cavalli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00300",
-    name: "Crea - Fumarola",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00301",
-    name: "Menegatti - Palmisani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00302",
-    name: "Alberti e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00303",
-    name: "Rocca - Leonetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00304",
-    name: "Diana - Mariani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00305",
-    name: "Betti - Castiglioni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00306",
-    name: "Tagliabue - Bertini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00307",
-    name: "Barbero - Murgia",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00308",
-    name: "Ferro, Bacci and Crespi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00309",
-    name: "Valentini e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00310",
-    name: "Picariello, Biagini and D'Alessandro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00311",
-    name: "Epifani SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00312",
-    name: "Amato, Kofler and Visentin",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00313",
-    name: "Innocenti, Tagliabue and Rapisarda",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00314",
-    name: "Laganà Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00315",
-    name: "Concas - Pepe",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00316",
-    name: "Babini, Cerrato and Morandi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00317",
-    name: "Pappalardo - Parrinello",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00318",
-    name: "Spano - Cingolani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00319",
-    name: "Carvelli - Agostinelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00320",
-    name: "Degano, Scarpa and Morelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00321",
-    name: "Manca Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00322",
-    name: "Magnani e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00323",
-    name: "Baraldi, Gruber and Masiero",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00324",
-    name: "Fabbri - Benedetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00325",
-    name: "Berto, Montemurro and Bartolini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00326",
-    name: "Lanza - Sbrana",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00327",
-    name: "Traverso - Franzè",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00328",
-    name: "Grieco - Schiavon",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00329",
-    name: "Canepa - Iezzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00330",
-    name: "Gallo Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00331",
-    name: "Carminati s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00332",
-    name: "Conte SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00333",
-    name: "Marini, Licitra and Zanetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00334",
-    name: "Bellini, Riggi and Bianchi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00335",
-    name: "De Giorgi SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00336",
-    name: "Salvadori e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00337",
-    name: "Greco - Betti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00338",
-    name: "Di Nardo, Bertolussi and Bertini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00339",
-    name: "Vailati - Bortot",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00340",
-    name: "Bressan Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00341",
-    name: "Lai s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00342",
-    name: "Di Iorio e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00343",
-    name: "Arena - Frigo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00344",
-    name: "Cannone, Benvenuti and Sorrentino",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00345",
-    name: "Franzoni, Milan and Martinelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00346",
-    name: "Epifani - Dal Farra",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00347",
-    name: "Pedrotti SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00348",
-    name: "Dalmasso - Caretti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00349",
-    name: "Guerra - Paolucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00350",
-    name: "Tanzi e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00351",
-    name: "Hofer - Rizzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00352",
-    name: "Biagini - Brambilla",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00353",
-    name: "Martinelli e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00354",
-    name: "Sbrana, Ricci and Epifani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00355",
-    name: "Valentini - Piazzolla",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00356",
-    name: "Esposito - Franco",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00357",
-    name: "Cossu s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00358",
-    name: "Pavan e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00359",
-    name: "Saponaro, Fiorucci and Pession",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00360",
-    name: "Russo, D'Orazio and De Marco",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00361",
-    name: "Gatti Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00362",
-    name: "Caputo, Gasser and Gabrielli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00363",
-    name: "Moretti, Carlucci and Messina",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00364",
-    name: "Parrinello, Giraudo and Moser",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00365",
-    name: "Rea - Acquistapace",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00366",
-    name: "Ferrari, Pircher and Pavan",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00367",
-    name: "Telesca s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00368",
-    name: "Licitra - Grillo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00369",
-    name: "Sbrana, Natale and Ferrari",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00370",
-    name: "Gori Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00371",
-    name: "Giacalone - Privitera",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00372",
-    name: "Di Matteo, Torchio and Paolucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00373",
-    name: "Picariello SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00374",
-    name: "Quinto, Picco and Carraro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00375",
-    name: "Bandini, Visintin and Pucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00376",
-    name: "Ercoli, Leonetti and Bruno",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00377",
-    name: "Veronese, Landi and Semeraro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00378",
-    name: "Falzone, Tesi and Zanetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00379",
-    name: "Grossi, Zennaro and Gori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00380",
-    name: "Rosset - Asaro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00381",
-    name: "Caputo - Valsecchi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00382",
-    name: "Gargiulo Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00383",
-    name: "Corti s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00384",
-    name: "Panzeri - Gregori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00385",
-    name: "Brumat, Cerrani and Mereu",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00386",
-    name: "De Luca - Traini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00387",
-    name: "Grossi - Pintus",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00388",
-    name: "Mazzoleno - Bertini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00389",
-    name: "Marega - Moffa",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00390",
-    name: "Corti - Giannini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00391",
-    name: "Ruggiero Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00392",
-    name: "Marini s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00393",
-    name: "Negri - Lucchesi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00394",
-    name: "Fiorucci - Rossi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00395",
-    name: "Zennaro SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00396",
-    name: "Musso, Huber and Melandri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00397",
-    name: "Redaelli e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00398",
-    name: "Fabbro, Penzo and Di Carlo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00399",
-    name: "Casadei - Zanella",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00400",
-    name: "Rinaldi - Castiglioni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00401",
-    name: "Di Dio, Nesti and Moretti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00402",
-    name: "Zanetta e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00403",
-    name: "Di Bari s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00404",
-    name: "Gamper, Pellegrini and Magnani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00405",
-    name: "D'Angelo - Crevatin",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00406",
-    name: "Parrinello - Torrisi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00407",
-    name: "Schiavon, Fusco and Morri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00408",
-    name: "Giovannini, Parodi and Martinelli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00409",
-    name: "Ferro e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00410",
-    name: "Guglielmi - Merola",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00411",
-    name: "Carlucci, Molteni and Marziali",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00412",
-    name: "Pozzo e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00413",
-    name: "Santarossa - Massaro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00414",
-    name: "Pivetta, Di Marco and Privitera",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00415",
-    name: "Tripodi, Gatti and Fabbro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00416",
-    name: "Gori e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00417",
-    name: "Sabia Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00418",
-    name: "Tumino, Di Dio and Vincenzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00419",
-    name: "Lenzi - Beltrame",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00420",
-    name: "Cecchi, Medeot and Nicoletti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00421",
-    name: "Cascone, Di Carlo and Vierin",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00422",
-    name: "Benedetti s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00423",
-    name: "Gamper - Palumbo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00424",
-    name: "D'Agostino, Ferrero and Romaniello",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00425",
-    name: "Berto, Bressan and Morri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00426",
-    name: "Fiore, Pieri and Serafini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00427",
-    name: "Bruzzone, Gabriele and Fiorini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00428",
-    name: "Di Marco Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00429",
-    name: "De Marco - Colucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00430",
-    name: "Palmisano, Natale and Molteni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00431",
-    name: "Favre - Sechi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00432",
-    name: "Di Marco SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00433",
-    name: "Ciuffreda - Beltrami",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00434",
-    name: "Santilli, Mariani and Spina",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00435",
-    name: "Bianchi - Licata",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00436",
-    name: "Manfredi - Greco",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00437",
-    name: "Ferro, Martino and Di Carlo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00438",
-    name: "Gasser, Sanfilippo and Lombardi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00439",
-    name: "Cossu - Santilli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00440",
-    name: "Zunino - Adami",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00441",
-    name: "Adami - Parodi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00442",
-    name: "Natale, Baldi and Sanfilippo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00443",
-    name: "Scarpa - Romagnoli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00444",
-    name: "Aiello - Beltrame",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00445",
-    name: "Zordan, Fiore and Belotti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00446",
-    name: "Mignogna Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00447",
-    name: "De Angelis e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00448",
-    name: "Magnani - Di Carlo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00449",
-    name: "Asaro - Mair",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00450",
-    name: "Acquadro, Sartor and Tiozzo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00451",
-    name: "Denti - Lanteri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00452",
-    name: "Bionaz, Proietti and Vichi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00453",
-    name: "Manfredi - Di Mauro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00454",
-    name: "Rocca, Belotti and Carrozzo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00455",
-    name: "Vercesi - Cavallari",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00456",
-    name: "Nicoletti - Moser",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00457",
-    name: "Penzo - Franzoni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00458",
-    name: "Capecchi SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00459",
-    name: "Tumino - Fumagalli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00460",
-    name: "Carrieri, Zunino and Bortoluzzi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00461",
-    name: "Montalbano - Fontana",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00462",
-    name: "Romagnoli, Panzeri and Moser",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00463",
-    name: "Montanari, Di Paolo and Bini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00464",
-    name: "Montalbano, Marangon and Fabbri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00465",
-    name: "Barbera - Proietti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00466",
-    name: "Cortinovis e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00467",
-    name: "Colombo Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00468",
-    name: "Landi s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00469",
-    name: "Liverani - Ballarin",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00470",
-    name: "Beltrami - Pepe",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00471",
-    name: "Carminati - Deiana",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00472",
-    name: "Rodigari, Di Bari and Leonetti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00473",
-    name: "Menna SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00474",
-    name: "Di Nardo SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00475",
-    name: "Bartolini s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00476",
-    name: "Gabriele - Neri",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00477",
-    name: "Corridori, Corti and Crepaldi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00478",
-    name: "Musso, Fiorini and Carletti",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00479",
-    name: "Matera, Musso and Baraldi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00480",
-    name: "Vincenzi e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00481",
-    name: "Marini Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00482",
-    name: "Ceccarelli, Sbrana and Amato",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00483",
-    name: "Palmisani, Spina and Andrisani",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00484",
-    name: "Epifani, Fabris and Fiorucci",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00485",
-    name: "Locatelli, Di Mauro and Usai",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00486",
-    name: "Granata, Zanella and Battistini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00487",
-    name: "D'Amico - Bartolini",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00488",
-    name: "Venezia - Scalise",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00489",
-    name: "Neri - Niccolai",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00490",
-    name: "Govoni, Iannaccone and Balboni",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00491",
-    name: "Ramella, Egger and Semeraro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00492",
-    name: "Liverani e figli",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00493",
-    name: "Conte Group",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00494",
-    name: "Musso, Tomasi and Ferraro",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00495",
-    name: "Traverso - Sartori",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00496",
-    name: "Kofler, Capecchi and Carrozzo",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00497",
-    name: "Huber - Bortolin",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00498",
-    name: "Palmisano SPA",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00499",
-    name: "Innocenti s.r.l.",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  },
-  {
-    abi: "00500",
-    name: "Angelini, Manca and Bertolussi",
-    logoUrl:
-      "https://upload.wikimedia.org/wikipedia/it/thumb/5/51/Intesa_Sanpaolo_logo.svg/320px-Intesa_Sanpaolo_logo.svg.png"
-  }
-];
-
-export const pans: ReadonlyArray<Card> = [
-  {
-    abi: "00001",
-    cardNumber: "1234",
-    cardPartialNumber: "11112222",
-    expiringDate: new Date("2021-10-07"),
-    productType: ProductTypeEnum.PP,
-    tokens: ["string"],
-    validityState: ValidityStateEnum.V
-  },
-  {
-    abi: "00002",
-    cardNumber: "5678",
-    cardPartialNumber: "33332222",
-    expiringDate: new Date("2021-10-07"),
-    productType: ProductTypeEnum.PP,
-    tokens: ["string"],
-    validityState: ValidityStateEnum.V
-  },
-  {
-    abi: "00003",
-    cardNumber: "5678",
-    cardPartialNumber: "33332222",
-    expiringDate: new Date("2021-10-07"),
-    productType: ProductTypeEnum.PP,
-    tokens: ["string"],
-    validityState: ValidityStateEnum.V
-  }
-];
